requirements.txt
atomicwrites==1.3.0
attrs==19.3.0
beautifulsoup4==4.8.1
boto3==1.10.28
botocore==1.13.28
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
colorama==0.4.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
defusedxml==0.6.0
Django==2.2.6
django-allauth==0.40.0
django-appconf==1.0.3
django-auth==0.1.9
django-avatar==4.1.0
django-bootstrap3==11.1.0
django-bootstrap4==1.0.1
django-cors-headers==3.2.0
django-developer-panel==0.1.2
django-extensions==3.0.9
django-filter==2.2.0
django-forms-bootstrap==3.0.0
django-guardian==2.1.0
django-hash-field==0.1
django-haystack==2.8.1
django-oauth-toolkit==1.2.0
django-polymorphic==2.1.2
django-rest-swagger==2.2.0
django-taggit==1.1.0
django-templated-mail==1.1.1
djangorestframework==3.10.3
djoser==2.0.3
docutils==0.15.2
gunicorn==20.0.4
idna==2.8
importlib-metadata==0.23
ipaddress==1.0.23
itypes==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
Markdown==3.1.1
MarkupSafe==1.1.1
more-itertools==7.2.0
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==19.2
Pillow==6.2.1
pluggy==0.13.1
psycopg2==2.8.4
py==1.8.0
pycparser==2.19
pycryptodome==3.9.4
PyJWT==1.7.1
pyotp==2.3.0
pyparsing==2.4.5
pytest==5.3.1
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.2.0
s3transfer==0.2.1
secret==0.8
simplejson==3.17.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
soupsieve==1.9.5
sqlparse==0.3.0
tabulate==0.8.6
uritemplate==3.0.0
urllib3==1.25.6
wcwidth==0.1.7
zipp==0.6.0
